+++ database/hash.go
@@ -614,6 +614,10 @@ func execHScan(db *DB, args [][]byte) redis.Reply {
 	if len(args) < 2 {
 		return &protocol.SyntaxErrReply{}
 	}
+	cursor, err := strconv.Atoi(string(args[1]))
+	if err != nil {
+		return protocol.MakeErrReply("ERR invalid cursor")
+	}
 	key := string(args[0])
 	dict, errReply := db.getAsDict(key)
 	if errReply != nil {
@@ -625,10 +629,6 @@ func execHScan(db *DB, args [][]byte) redis.Reply {
 	count, pattern, optErr := parseTypeScanOptions(args, 2)
 	if optErr != nil {
 		return optErr
-	}
-	cursor, err := strconv.Atoi(string(args[1]))
-	if err != nil {
-		return protocol.MakeErrReply("ERR invalid cursor")
 	}
 
 	keysReply, nextCursor := dict.DictScan(cursor, count, pattern)