+++ database/set.go
@@ -413,6 +413,10 @@ func stringsToSetReply(members []string) redis.Reply {
 }
 
 func execSScan(db *DB, args [][]byte) redis.Reply {
+	cursor, err := strconv.Atoi(string(args[1]))
+	if err != nil {
+		return protocol.MakeErrReply("ERR invalid cursor")
+	}
 	key := string(args[0])
 	set, errReply := db.getAsSet(key)
 	if errReply != nil {
@@ -424,10 +428,6 @@ func execSScan(db *DB, args [][]byte) redis.Reply {
 	count, pattern, optErr := parseTypeScanOptions(args, 2)
 	if optErr != nil {
 		return optErr
-	}
-	cursor, err := strconv.Atoi(string(args[1]))
-	if err != nil {
-		return protocol.MakeErrReply("ERR invalid cursor")
 	}
 
 	keysReply, nextCursor := set.SetScan(cursor, count, pattern)